backend/requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0